03-amazon-reviews-project/ingestion/requirements.txt
snowflake-connector-python==3.12.2
pandas==2.2.2
python-dotenv==1.0.1
tqdm==4.66.4